add : created a File-to-Bytes Reader(PNG)
Opens the file sample-png-file.png.
Reads all the binary data from it into a byte[].
Prints each byte as a signed integer (-128 to 127).
Example: If a byte in file is 0xFF (255 unsigned), Java prints it as -1 because Java bytes are signed.
Prints the total number of bytes (file size in bytes).